Refactor post#index view
The _post partial had too much logic in order to decide
if it was going to render as a small card or a individual post.

In order to split both cases the posts#index card rendering
was splitted using two partials.
- posts/_card
- comments/_form

The intent is to use the comments/_form partial in the
posts#show view