Add eager loading to posts#index
The posts#index and #show views where making too many
SQL calls, mainly because of the posts.comments.count calls,
but also for the post.user in the index.

Also replaced the post.comments #count call for a #.size call.
The #count method always results in a SQL count() call, but
#size uses the already fetched Association.