Fix bad add_username migration
Adding a UNIQUE NOT NULL column isn't an easy task.
Many SQL database implementations won't allow to add a NOT NULL
column without a DEFAULT value, but if UNIQUE is also added to the mix
it simply won't be able. To acomplish the task 3 steps had to be taken:

  1. create the column without NOT NULL constrain
  2. populate the new column
  3. modify the column to add the constrain

I opted for a raw SQL call inside the migration because doing it with
Ruby could lead to more issues in the future as mentioned in this blog post:

https://railsguides.net/change-data-in-migrations-like-a-boss/

The SQL template was found in StackOverflow (no surprises here)

https://stackoverflow.com/questions/512451/
how-can-i-add-a-column-that-doesnt-allow-nulls-in-a-postgresql-database