Add FriendRequest validations
Add database validation for unique index to avoid duplicated
FriendRequests.

Add new validator to the model to verify the same condition.

The `on: :create` option is necessary or an status update will raise an
error complaining that an identical requests already exists, altough is
refering to the updated requests itself.